Convert the UI reducer/selector/actions to typescript and add handling for the confirmationModelActive key.
