--- src/ui/actions.js
@@ -1,14 +1,0 @@
-// Action labels
-export const actionsList = {
-  userMenuActive: "TOGGLE_USER_MENU",
-};
-
-/**
-  Persist user menu visibility between page renders
-*/
-export function userMenuActive(toggle) {
-  return {
-    type: actionsList.userMenuActive,
-    payload: toggle,
-  };
-}
